+++ main.py
@@ -7,6 +7,7 @@ from src.presentation.routes.topics_route import app_topics_router
 from src.presentation.routes.input_route import input_router
 from src.presentation.routes.context_route import context_router
 from src.presentation.routes.pdf_route import pdf_router
+from src.presentation.routes.search_route import search_router
 from src.core.settings import app_settings
 
 settings = app_settings()
@@ -56,6 +57,11 @@ app.include_router(
 app.include_router(
     app_topics_router,
     tags=["app_topics"]
+)
+
+app.include_router(
+    search_router,
+    tags=["search"]
 )
 
 app.include_router(