+++ CHANGELOG.md
@@ -47,6 +47,7 @@ All notable changes to this project will be documented in this file.
 - *(id)* Added a pattern for identification of the cards
 - *(ai)* Improved to clean useless char data for input
 - *(ai)* Improved to have amount of question and language of route to generate question
+- *(adapter)* Created a search method
 
 ### ⚙️ Miscellaneous Tasks